backend/requirements.txt
arrow==1.3.0
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
autopep8==1.6.0
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
constantly==15.1.0
cryptography==41.0.3
daphne==4.0.0
dateparser==1.2.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.4
django-channels-jwt-auth-middleware==1.0.0
django-ckeditor==6.7.1
django-ckeditor-5==0.2.12
django-cors-headers==3.11.0
django-crispy-forms==2.1
django-import-export==3.3.7
django-jazzmin==2.6.1
django-js-asset==2.2.0
django-mathfilters==1.0.0
django-taggit==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
et-xmlfile==1.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
MarkupPy==1.14
moment==0.12.1
msgpack==1.0.5
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.2.0
python-dateutil==2.9.0.post0
pytz==2023.3
PyYAML==6.0.1
redis==5.0.0
regex==2023.12.25
service-identity==23.1.0
shortuuid==1.0.13
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
times==0.7
toml==0.10.2
Twisted==23.8.0
txaio==23.1.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.7.1
tzdata==2021.5
tzlocal==5.2
xlrd==2.0.1
xlwt==1.3.0
zope.interface==6.0
